The "Smooth Movement" Update
In this update I added smooth movement.

Rather than waiting for the browser to register multiple key presses
(you press a key, there is a delay and then the key is spammed), the
program now stores if the key is currently pressed down or not. Each
frame, pending if the key is pressed, the players position will update.

This update also adds support for multiple keys to be pressed and
detected at once. Now the player can move forward and side strafe .